pyproject.toml
[project]
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary",
    "azure-storage-blob",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
